requirements.txt
runpod
boto3
requests
pillow
torch==2.8.0
torchvision==0.23.0
torchaudio==2.8.0
diffusers
transformers
accelerate
safetensors
peft
scipy
# Note: Flash attention remains in the bootstrap because it's hardware-specific
